requirements.txt
watchdog==0.8.3
websocket-client==0.40.0
coverage==4.3.4
Django==1.11
mock==2.0.0
nose2==0.6.5
